Lab3: rename app.
